Azure: skip backend pool if attached to an outbound rule
When configuring egress IP on Azure, omit the backend address pool if it's attached to outbound rule, since attaching a secondary IP is not allowed in such cases on Azure, resulting in an OutboundRuleCannotBeUsedWithBackendAddressPoolThatIsReferencedBySecondaryIpConfigs error.

This affects the following setups:
- azure private clusters
- aro private clusters
- aro public custers

WARNING: as a result, the egress IP will have no outward connectivity. Further investigation is needed to let the egress IP communicate to the outside.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>